host_sections: make MultiHostSections a member of _legacy_mode
