+++ cmk/base/checking/_legacy_mode.py
@@ -3,12 +3,15 @@
 # Copyright (C) 2019 tribe29 GmbH - License: GNU General Public License v2
 # This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
 # conditions defined in the file COPYING, which is part of this source code package.
-"""Performing the actual checks."""
+"""Legacy version of running checks on clusters"""
 
+import sys
 from typing import (
     Any,
     AnyStr,
     cast,
+    Callable,
+    Dict,
     Iterable,
     List,
     Optional,
@@ -16,10 +19,13 @@ from typing import (
     Union,
 )
 
+import cmk.utils.caching
 import cmk.utils.debug
-from cmk.utils.exceptions import MKGeneralException, MKTimeout
+from cmk.utils.check_utils import section_name_of
+from cmk.utils.exceptions import MKGeneralException, MKTimeout, MKParseFunctionError
 from cmk.utils.type_defs import (
     HostAddress,
+    HostKey,
     HostName,
     MetricTuple,
     SectionName,
@@ -39,9 +45,15 @@ import cmk.base.ip_lookup as ip_lookup
 import cmk.base.item_state as item_state
 import cmk.base.utils
 
-from cmk.base.check_api_utils import MGMT_ONLY as LEGACY_MGMT_ONLY
+from cmk.base.api.agent_based import register as agent_based_register
+from cmk.base.check_api_utils import (
+    MGMT_ONLY as LEGACY_MGMT_ONLY,
+    HOST_PRECEDENCE as LEGACY_HOST_PRECEDENCE,
+)
 from cmk.base.check_utils import LegacyCheckParameters, Service
-from cmk.base.sources.host_sections import HostKey, MultiHostSections
+from cmk.base.sources.host_sections import ParsedSectionsBroker, ParsedSectionContent
+
+from cmk.core_helpers.cache import ABCRawDataSection
 
 from .utils import (
     AggregatedResult,
@@ -54,7 +66,7 @@ ServiceCheckResultWithOptionalDetails = Tuple[ServiceState, ServiceDetails, List
 
 
 def get_aggregated_result(
-    multi_host_sections: MultiHostSections,
+    parsed_sections_broker: ParsedSectionsBroker,
     hostname: HostName,
     ipaddress: Optional[HostAddress],
     service: Service,
@@ -82,6 +94,7 @@ def get_aggregated_result(
     section_name = legacy_check_plugin_name.split('.')[0]
 
     section_content = None
+    multi_host_sections = _MultiHostSections(parsed_sections_broker)
     mgmt_board_info = config.get_management_board_precedence(section_name, config.check_info)
     source_type = SourceType.MANAGEMENT if mgmt_board_info == LEGACY_MGMT_ONLY else SourceType.HOST
     try:
@@ -229,3 +242,175 @@ def _sanitize_check_result_infotext(infotext: Optional[AnyStr],
         return infotext.decode('utf-8')
 
     return infotext
+
+
+class _MultiHostSections:
+    """Container object for wrapping the host sections of a host being processed
+    or multiple hosts when a cluster is processed. Also holds the functionality for
+    merging these information together for a check"""
+    def __init__(self, parsed_sections_broker: ParsedSectionsBroker) -> None:
+        super().__init__()
+        self._parsed_sections_broker = parsed_sections_broker
+        self._section_content_cache = cmk.utils.caching.DictCache()
+
+    def get_section_content(
+        self,
+        host_key: HostKey,
+        management_board_info: str,
+        check_plugin_name: str,
+        for_discovery: bool,
+        *,
+        cluster_node_keys: Optional[List[HostKey]] = None,
+        check_legacy_info: Dict[str, Dict[str, Any]],
+    ) -> Union[None, ParsedSectionContent, List[ParsedSectionContent]]:
+        """Prepares the section_content construct for a Check_MK check on ANY host
+
+        The section_content construct is then handed over to the check, inventory or
+        discovery functions for doing their work.
+
+        If the host is a cluster, the sections from all its nodes is merged together
+        here. Optionally the node info is added to the nodes section content.
+
+        It handles the whole data and cares about these aspects:
+
+        a) Extract the section_content for the given check_plugin_name
+        b) Adds node_info to the section_content (if check asks for this)
+        c) Applies the parse function (if check has some)
+
+        It can return an section_content construct or None when there is no section content
+        for this check available.
+        """
+
+        section_name = section_name_of(check_plugin_name)
+        cache_key = (host_key, management_board_info, section_name, for_discovery,
+                     bool(cluster_node_keys))
+
+        try:
+            return self._section_content_cache[cache_key]
+        except KeyError:
+            pass
+
+        section_content = self._get_section_content(
+            host_key._replace(source_type=SourceType.MANAGEMENT if management_board_info ==
+                              LEGACY_MGMT_ONLY else SourceType.HOST),
+            check_plugin_name,
+            SectionName(section_name),
+            for_discovery,
+            cluster_node_keys=cluster_node_keys,
+            check_legacy_info=check_legacy_info,
+        )
+
+        # If we found nothing, see if we must check the management board:
+        if (section_content is None and host_key.source_type is SourceType.HOST and
+                management_board_info == LEGACY_HOST_PRECEDENCE):
+            section_content = self._get_section_content(
+                host_key._replace(source_type=SourceType.MANAGEMENT),
+                check_plugin_name,
+                SectionName(section_name),
+                for_discovery,
+                cluster_node_keys=cluster_node_keys,
+                check_legacy_info=check_legacy_info,
+            )
+
+        self._section_content_cache[cache_key] = section_content
+        return section_content
+
+    def _get_section_content(
+        self,
+        host_key: HostKey,
+        check_plugin_name: str,
+        section_name: SectionName,
+        for_discovery: bool,
+        *,
+        cluster_node_keys: Optional[List[HostKey]] = None,
+        check_legacy_info: Dict[str, Dict[str, Any]]
+    ) -> Union[None, ParsedSectionContent, List[ParsedSectionContent]]:
+        # Now get the section_content from the required hosts and merge them together to
+        # a single section_content. For each host optionally add the node info.
+        section_content: Optional[ABCRawDataSection] = None
+        for node_key in cluster_node_keys or [host_key]:
+
+            try:
+                host_section_content = self._parsed_sections_broker[node_key].sections[section_name]
+            except KeyError:
+                continue
+
+            if section_content is None:
+                section_content = host_section_content[:]
+            else:
+                section_content += host_section_content
+
+        if section_content is None:
+            return None
+
+        assert isinstance(section_content, list)
+
+        return self._update_with_parse_function(
+            section_content,
+            section_name,
+            check_legacy_info,
+        )
+
+    @staticmethod
+    def _update_with_parse_function(
+        section_content: ABCRawDataSection,
+        section_name: SectionName,
+        check_legacy_info: Dict[str, Dict[str, Any]],
+    ) -> ParsedSectionContent:
+        """Transform the section_content using the defined parse functions.
+
+        Some checks define a parse function that is used to transform the section_content
+        somehow. It is applied by this function.
+
+        Please note that this is not a check/subcheck individual setting. This option is related
+        to the agent section.
+
+        All exceptions raised by the parse function will be catched and re-raised as
+        MKParseFunctionError() exceptions."""
+        # We can use the migrated section: we refuse to migrate sections with
+        # "'node_info'=True", so the auto-migrated ones will keep working.
+        # This function will never be called on checks programmed against the new
+        # API (or migrated manually)
+        if not agent_based_register.is_registered_section_plugin(section_name):
+            # use legacy parse function for unmigrated sections
+            parse_function = check_legacy_info.get(str(section_name), {}).get("parse_function")
+        else:
+            section_plugin = agent_based_register.get_section_plugin(section_name)
+            parse_function = cast(Callable[[ABCRawDataSection], ParsedSectionContent],
+                                  section_plugin.parse_function)
+
+        if parse_function is None:
+            return section_content
+
+        # (mo): ValueStores (formally Item state) need to be *only* available
+        # from within the check function, nowhere else.
+        orig_item_state_prefix = item_state.get_item_state_prefix()
+        try:
+            item_state.set_item_state_prefix(section_name, None)
+            return parse_function(section_content)
+
+        except item_state.MKCounterWrapped:
+            raise
+
+        except Exception:
+            if cmk.utils.debug.enabled():
+                raise
+            raise MKParseFunctionError(*sys.exc_info())
+
+        finally:
+            item_state.set_item_state_prefix(*orig_item_state_prefix)
+
+    def legacy_determine_cache_info(self, section_name: SectionName) -> Optional[Tuple[int, int]]:
+        """Aggregate information about the age of the data in the agent sections
+
+        This is in checkers.g_agent_cache_info. For clusters we use the oldest
+        of the timestamps, of course.
+        """
+        cache_infos = [
+            host_sections.cache_info[section_name]
+            for host_sections in self._parsed_sections_broker.values()
+            if section_name in host_sections.cache_info
+        ]
+
+        return (min(at for at, _ in cache_infos),
+                max(interval for _, interval in cache_infos)) if cache_infos else None
